+++ lib/infinity_one_web/router.ex
@@ -21,6 +21,7 @@ defmodule InfinityOneWeb.Router do
     plug(:put_secure_browser_headers)
     plug(InfinityOne.Plugs.Setup)
     plug(Coherence.Authentication.Session, protected: true)
+    plug(InfinityOne.Plug.SingleLogin)
   end
 
   pipeline :api do