+++ plugins/one_chat/lib/one_chat_web/channels/user_channel.ex
@@ -750,6 +750,7 @@ defmodule OneChatWeb.UserChannel do
       :delete_subscription
     subscribe_callback "user:" <> user_id, "room:update",
       :room_update
+    subscribe_callback "user:" <> user_id, "logout", :logout
     subscribe_callback "user:" <> user_id, "webrtc:offer", :webrtc_offer
     subscribe_callback "user:" <> user_id, "webrtc:answer", {WebrtcChannel, :webrtc_answer}
     subscribe_callback "user:" <> user_id, "webrtc:leave", {WebrtcChannel, :webrtc_leave}
@@ -1417,6 +1418,19 @@ defmodule OneChatWeb.UserChannel do
         .prepend('<span class="unread">#{unread}</span>');
       """
     push socket, "update:alerts", %{}
+  end
+
+  def logout(_event, payload, socket) do
+    key = Rebel.Core.exec_js!(socket, ~s/window.ucxchat.key/)
+    if key == payload.creds do
+      Client.toastr(socket, :warning,
+        ~g(Logging you out. Someone logged into your account from another device.))
+      spawn fn ->
+        Process.sleep(3_000)
+        Rebel.Core.async_js(socket, ~s(window.location.href="/logout"))
+      end
+    end
+    socket
   end
 
   def room_update(_event, payload, socket) do